requirements.txt
django==4.0.4
djangorestframework==3.13.1
drf-yasg
django-cors-headers
psycopg2-binary
PyJWT